feat: allow for skill in roll formulas (#753)
Previously one could not use skills in roll formulas.
This change allows for using skills using @skills.SkillName.
Spaces and other special charachters are remove from skill names, 
thus Gun Combat becomes GunCombat. See the item page for the specific name.
Normally the value of the untrained skill value is returned when an actor
lacks the skill. Adding an undnerscore (_) at the end of the name will make the
skill default to 0 if it doesn't exist or for a negative value. The above example
would thus be @skills.GunCombat_ to return 0 if the actor lacks the skill.

By @jonepatr and feedback from @marvin9257