feat: Ship positions (#801)
New ship positions are added where one can create those ship positions
dynamically and drag/drop actors on them.
Actors can be moved around between ship positions by dragging and
dropping (or if moved outside a ship position removed from the ship).
One actor can only be at one position at a time. If there is only one
actor at a position it will be assumed that actor performed the action
when it is clicked, if there are multiple actors, one needs to click on
an actor to mark it, and then use the action.
Each position can have multiple actions. Currently three actions are
supported, sending chat message, doing a skill roll, or firing energy
weapons (the fire energy weapon isn't fully fleshed out yet. more
actions will come). A quick way of adding skill roll actions is to drag
and drop a skill onto the ship position item sheet or onto the ship
position in the ship sheet, it will automatically create an action with
that skill roll.
For the skill roll you can write four things as the "command", we will
exemplifying with the skill "Gun Combat";
"Gun Combat", "Gun Combat/STR", "Gun Combat 10+", or "Gun Combat/STR
10+". Adding a characteristic will override the default characteristic
for that skill, and adding a difficulty will set the difficulty of the
skill check.
Components are shown as a reference for including into the command. For
example fire energy weapons would take a weapon component. This is not
fully used yet however (more than saying which component was used to
shoot with).
Reordering of the ship positions is done by going in and editing the
order number on the ship position item.

Fix #429

By @jonepatr with a lot of help from @marvin9257