feat: Create chat buttons (#1247)
This release introduces follow-up roll buttons on the chat message. 
For all rolls, the information table is toggled with a "question-mark" button.
For weapons rolls without auto damage - a "person-burst" button will roll damage for that check.
For skill rolls, chain and opposed roll buttons appear. A dialog pops-up asking for a selected skill.  Note that the "assisting" roll comes first. As a GM, you must have an actor token selected to do a follow-on roll.  As a player, you must own a actor on the scene.

Some bug fixes: 
* Prevent active effects drag and drop (thanks TImb)
* Set max drive value to 15 rather than 10 (thanks Debinani)