feat: Computer attachment changes (#1274)
This update adds the ability to toggle software attachments on computer items active and inactive from the actor sheet.  A display of used and available bandwidth is also listed under "short description" depending on software that is active. Note that the software "consumable" must be set as an attachment to work.  Suggestions per @HazardSW.  

Also fixed is a typo causing chain bonuses to not be applied.