fix: temp fixes to encumbrance indicators (#1341)
This update fixes a bug with encumbrance indicators in v11 when dropping an item onto a linked actor sheet.  The encumbrance indicator should update based on carry weight.  However with this workaround, the user can not delete the effect from effects tab, but it can be disabled.  Thanks to @sharcashmo for finding.