fix: missing untrained skills on actor (#1368)
This update adds a check for actors missing the "Untrained" skill when a world loads.  If the skill missing, the system will add the skill to the actor.  Note that this skill must be present even if not visible (as selected by settings).  Thanks to @ForjaSalvaje for finding and helping to debug this problem.

This update also fixes an issue with the Skills List on the actor sheet displaying a horizontal scroll bar when the list is longer than one screen.  Thanks to @Surge for finding.