feat: add degrees of success (#1435)
This release includes a feature to add degrees of success descriptions to non-attack rolls.  There are three options for the new ruleset setting ("Show degrees of success for non-attack rolls"): 
* None: don't display degrees of success
* CE: use Cepheus Engine scale
* Other: expand CE scale to include marginal success/failure.
Each ruleset choice now include a default setting for this feature, as well.  For those configuring for MgT2, the "Other" setting is the closest choice.