fix: allow item-piles module to work with damage drop (#1541)
This small update allows the item piles module to work with dragging damage cards onto tokens.  Thanks to the suggestion of @Silent Phoenix.
Please note that you will have to enable (or re-enable) the experimental feature that allows dragging and dropping on tokens.  Previously, this option was disabled if the item piles module was active.