feat: Add in-world time integration to timeframe rolls (#1686)
With this release, Twodsix can optionally change the in-world time based on the results of a timeframe roll.  This feature integrates with modules Simple Calendar, SmallTime, or any module that uses the FVTT internal game clock. See the wiki for more information: https://github.com/xdy/twodsix-foundryvtt/wiki/Tracking-Time

Thanks to WCV Consensus / nearlyNonexistent for the suggestion.