feat: move to v13 (#1667)
This update migrates Twodsix to the Foundry V13 baseline.  This version moves all sheets, macros, and objects to the AppV2 format.  This change required a significant refactor of the code.  So there are bound to be bugs.  Please test on a trial world before using for a live game.  

Other major changes:
* AP values may be formulas
* Limited ability to drag-drop Active Effects between actors and Journal Entries
* Allow ability to drag-drop item folders onto actors
* Try to make skill names unique by appending number if needed (this is so that skills proxy works correctly)
* Add @ItemList enricher to drag multiple items from items sidebar
* Allow EDU, SOC and INT characteristics to be hidden if zero.
* Support secrets for prose-mirror sheet areas
* Add better scrollbar and font scaling to ship sheet
* Add a roll setting option to do xd6 style rolls (e.g. Nomad)
* Items may have more than one pdf pager reference
* Allow tokens in AOE attack to be auto targeted
* Add move ruler coloring based on FVTT internal method
* Misc bug fixes

BREAKING CHANGE: This release moves to the FVTT v13 baseline.  It requires a ONE WAY migration to the new format.  You won't be able to go back to v12 once your world migrates.  Also, chat message buttons won't work for chat cards that are pre v13.  PLEASE MAKE A BACKUP!